Moved the looping functionality for mm_nls into a new function, nls_loop
This function handles the result data (substrate concentration, initial
rate, min and max time, and dataset exclusion), and allows the user to
reprocess or exclude data.  It calls regression_loop and mm_nls when
appropriate, and ultimately returns the result data and the nls
regression object.

The pair of mm_nls and nls_loop is analogous to make_reggraph and
regression_loop.