Added mm_validation function
Added a function to validate the input raw data and result_list data.
Testing seems to show that it is working.  The function will stop
execution if:
1. Non-numeric data is provided anywhere except in the concentration
dataset names.
2. The result_list concentration datasets do not match the raw_data
concentration datasets.
3. The result_list file is not in the correct format.

If the concentration names do match, but are not numeric, the program
will warn the user and prompt if they want to continue.

If result_list is not provided or the concentration data series names
are not numeric, returns FALSE.  Otherwise, returns TRUE (ie. we can do
NLS analysis).